Make PatriciaTreeSet fulfill C++ container concepts
Summary:
I was writing some unit tests recently with PatriciaTreeSets and noted
that I wasn't able to use the gtest UnorderedElementsAre matcher on it due to it
not defining certain standard container member types. This diff uses the
`BOOST_CONCEPT_ASSERT` mechanism to make sure that we have a comprehensive
implementation of the C++ container / iterator types and won't stumble on
similar issues in the future.

Reviewed By: thezhangwei

Differential Revision: D13616206

fbshipit-source-id: 8653df0e05d4983d83eb2ee33f2040dc28baf245