add unit tests that expose a bug in the patricia tree leq operator
Summary: `PatriciaTreeMapAbstractPartition` − which is based on `PatriciaTreeMap` with a default value of bottom − has an unsound less-or-equal operator. It does not properly compare tree nodes when the left hand side is not included in the right hand side. This patch adds tests to demonstrate the bug. A follow-up will fix it.

Reviewed By: int3

Differential Revision: D22992055

fbshipit-source-id: 7c455b73f578626f1773f059f591da6c8f792282