implement a set on top of a sorted vector
Summary: This patch implements an unordered set using a sorted vector. It is similar to `boost::container::flat_set` but provides the same interface as `PatriciaTreeSet`. This can be used as a performance optimization when handling small sets. This will be used in a follow-up to implement a small topped set abstract domain.

Reviewed By: yuxuanchen1997

Differential Revision: D24747018

fbshipit-source-id: 58a0329644fff203bca2d827cdd0c813ae940ec5