Fixing monotonic fixpoint iterators
Summary:
{gif:1wv9tbwx}

Without the change, the events that can (and have been observed to) play out in practice are like this:
- Resetting counters to NumOuterSchedPreds happens in some order. In particular, at some point the head counter is reset, which then can immediately kick off further work concurrently.
  - This concurrent further work can and will increment counters of inner nodes.
- Eventually, the ongoing resetting logic will also reset the counter of the inner node, and (the way the code used to be) resets it fully, throwing away the intermediate (and legitimate) increments! This will effectively cut off the computation, and we won't actually proceed to compute the actual fixpoint.

The paper (https://arxiv.org/pdf/1909.05951.pdf) says "Operations on N are atomic". If that is to mean performing the (resetting) bulk-updates as a single atomic operation, then the paper is correct, and it's just that the implementation was slightly wrong, as it handled atomicity point-wise. In any case, I believe, that my change addresses the issue elegantly while sticking to the efficient point-wise atomic updates (together with the setting the self-counter to 0 before processing the component and incrementing other component counters).

The paper authors might want to clarify what they mean for future readers...

Reviewed By: minjang, ssj933

Differential Revision: D26033374

fbshipit-source-id: f53fdacc21211f0192bc1ca71731b9ea8f5562ef