Fix up PatriciaTree iterators.
Summary:
An upcoming diff merges these two iterator implementations into one. However, set's iterator does not quite comply with the standard forward iterator requirements; this makes merging the implementations awkward, due to special casing.

This is easy to fix, but I wanted to call it out explicitly so it isn't a subtle change in a larger diff.

Reviewed By: yuxuanchen1997

Differential Revision: D37281001

fbshipit-source-id: 3d6229d8ea225effb189245d56bbd31e3d41aa37