Remove unnecessary ref counting.
Summary:
Every time we `static_pointer_cast`, we are creating a fresh `intrusive_ptr` that increments and later decrements the node ref count. Most of the time, we don't need an intrusive pointer and just need to static cast the pointer.

This diff adds `as_leaf` and `as_branch` which will conditionally perform this cast (else return nullptr) for convenience. Then applies it throughout the implementation to remove ref counting.

The only places where it is necessary is when we do an operation that actually generates a new intrusive node pointer, in `update_leaf`. As such, there are no more `static_pointer_cast`s to a branch node and all of the leaf updates are modified to move around intrusive node pointers rather than copy.

Reviewed By: NTillmann

Differential Revision: D37293703

fbshipit-source-id: f224b8a904a5c1b43c87580d6bf6fd55a0192bce