Minor shuffle of BitVectorLattice.
Summary:
The upcoming diffs add a lot more implementation inside the `fad_impl` namespace. This ends up putting the actual useful public class, `BitVectorLattice`, way down at the bottom and hard to find.

This rearranges so that the public interface goes first.

 ---

The goal of this stack is to speed up BitVectorLattice encode & decode operations by ~6x-16x. The first three diffs are some refactoring and restriction so the implementation details are less entangled with their use. Then the performance improvements are added.

Reviewed By: yuxuanchen1997

Differential Revision: D37775182

fbshipit-source-id: 94eebba287282bab3a7589e155d6c37e646712e2