Refactor tree operation, implement remove
Summary:
Make it more like the C++ version. One shared template to perform everything.
The `op` will be called only once. If the node exists, the argument will be `Some`. If the node doesn't exist, the argument will be passed as `None`. Then replaces the value with return value of `op`. If `op` returned `None`, the matching value will be removed.

Differential Revision: D38929715

fbshipit-source-id: 79650e9b34a1095efe536185f6206a4e6e775862