Add derive macros for DisjointUnionAbstractDomain
Summary:
Derive macros for DisjointUnionAbstractDomain from an enum of abstract domains.

Rust does not allow variadic template arguments, which is probably the right design decision. In order for us to still be able to compose abstract domains together, we can implement it as a `#[derive(...)]` macro.

The disjoint union abstract domain is bottom/top whenever the underlying variant is at bottom/top, regardless of which variant we use. This potentially causes a un-normalizable representation of top and bottoms. The direct ramification to it is that we probably shouldn't directly derive `Eq` on enums that derive `DisjointUnion`.

This isn't yet checking whether the enum variants satisfy the requirement of having only one field which is an abstract domain. It will error out, but the error message will be a bit nonsensical.

Differential Revision: D38993647

fbshipit-source-id: 18761884a84e0494b9d6010b461a6499da77c5b0