implement HashMapAbstractEnvironment
Summary:
Useful for modelling variable values. Should be equivalent to `HashedAbstractEnvironment` in C++. Top values are implicitly stored.

It's sort of like an inverted domain for HashMapAbstractPartition. But in reality, Environment has one more semantic difference from an inverted partition. When any of its member mapped domain is assigned Bottom, the entire environment falls to bottom. This doesn't happen for a partition when it's assigned top.

Differential Revision: D39192065

fbshipit-source-id: e0c0927b2b9139711856e0654f29893bb4af8a00