Impl SetAbstractDomainOps for PatriciaTreeSet
Summary:
Just realized that SetAbstractDomainOps is not implemented for PatriciaTreeSet. This makes it cannot be used exchangeably with HashSetAbstractDomain.

The main issues of implementing SetAbstractDomainOps for PatriciaTreeSet is that `iter()` of PatriciaTreeSet returns an iterator of values instead of references, which is incompatible with the current signature of `elements()`.

To address this, we choose to rely on into_iter() of HashSet/PatriciaTreeSet, which can return either reference or value since PowersetLattice is a generic type. Note that we can not use into_iter to implement elements() due to the lifetime bound on the impl of IntoIterator for PatriciaTreeSet. So we remove `elements()` from `SetAbstractDomainOps`.

Reviewed By: yuxuanchen1997

Differential Revision: D39605281

fbshipit-source-id: ba0e7074da10a4a737a2a71a2b71a46cfdbecafa