Require abstract domain classes to be final
Summary:
In C++, one can mark a class as `final`, which prevents it from being extended.

This allows compiler optimizations, most notably when calling a virtual method on a `final` class: the compiler knows that it cannot call a method of a child class, hence it can turn the virtual call into a static call.

In general, we should avoid using inheritance in C++, except for polymorphism (when we need to see something as a generic base class) or when using CRTP. If one needs to encapsulate a class, writing a wrapper class with the wrapped class as a field is recommended.

In conclusion, abstract domain implementations should always be marked as `final`. Let's enforce it with a `static_assert`.

Reviewed By: arnaudvenet

Differential Revision: D45239210

fbshipit-source-id: b744d336c50c2c0966131547025151a92ded0d1a