Implement visitor methods on most data structures
Summary:
I have noticed that patricia tree iterators internally use a stack, which requires allocating memory on the heap.
This can be detrimental to performance when using heavily nested maps, and iterating on the whole structure. This leads to a lot of allocations and deallocations.
To prevent this problem, this diff implements `visit` methods on most data structures.
`visit` uses recursive functions, which means it never allocate memory.

Reviewed By: arnaudvenet

Differential Revision: D49922681

fbshipit-source-id: 00de545e61e17dfb2b87501d0c9c943c2e5f7cf5