Add a macro to mark unused variables
Summary:
When compiling with both `-Wunused-variable` and `-DNDEBUG`, we are getting errors due to unused variables.
This is because we have variables that are created and only used in assertions.

Let's introduce a `SPARTA_UNUSED` macro that does a static cast to void, which should get optimized away.
Let's also use this to introduce `SPARTA_ASSERT` which is just an alias for the C assert, for now.

Reviewed By: GerbenJavado

Differential Revision: D51588387

fbshipit-source-id: f8ccc03bace8112f6b26fdad850c5fad47f388c4