Move the flags class into sparta
Summary:
In Mariana Trench, we have a convenient utility class called `Flags` which implements a bitset on top of an enum where values are powers of 2. This makes it very efficient to store a small set of flags.

I would like to re-use this in other tools (namely, Angliru). My two options are:
* Copy/paste the code of that class.
* Move this in a library that "other" tools can/will use. This is what this diff does.

The downside of course is that sparta might become a dumping group for shared code, which is not what we want.
I believe this is small and generic enough that it could fit in sparta.

Reviewed By: anwesht

Differential Revision: D51589404

fbshipit-source-id: 87ad709a8b786126caa48e31fe58f797d2415bca