Fix install interface to use CMAKE_INSTALL_INCLUDEDIR
Currently, the sparta target will use `$CMAKE_INSTALL_PREFIX/include` as
the include directory when installed. If the user sets a custom
CMAKE_INSTALL_INCLUDEDIR, however, then this will not be the correct
location. This commit changes the include directory of the sparta target
to be consistent with the actual installed location.