Take the less-equal binary function as a argument in PatriciaTreeCore::leq
Summary:
We want to use patricia trees in another tool internally, to improve performance.
That tool has an abstract domain interface that has an additional parameter for its `leq` operator.
Unfortunately, the `leq` operator in sparta is a static method in the `ValueInterface`, which cannot be adapted for that tool.

To unblock the situation, let's make `PatriciaTreeCore` (the internal class representing patricia tree) accept an arbitrary operator for its `leq` method.
Note that `PatriciaTreeMap`, `PatriciaTreeMapAbstractEnvironment` and `PatriciaTreeMapAbstractPartition` are unchanged.

We also expose a `find` method that returns nullptr when the key is not found. This will also be necessary.

Reviewed By: arnaudvenet

Differential Revision: D60052170

fbshipit-source-id: c27eeaef4872fe8d4b9b93409e5017622faf1952